+++ CHANGELOG.md
@@ -15,6 +15,8 @@
   - always encode DateTime with time zone
 - HttpClient
   - HTTP/2.0 client can reconnect if connection was lost
+- Persistence
+  - SubQuery as new QuerySource type
 
 ## 0.16.5
 - TransferCodec, TransferBean